Adding the Gyro for Field-Oriented Drive
so we are high-key small brain and forgot to add the gyro stuff to the robot. All you have to do is add the gyro angle at the end of the driveCartesian method to make it do the field oriented drive things. Side note, we the gyro with all the weird letters and numbers, and that only takes the SPI.findthingport thing, not an integer.